At the top of each page please make the ‘Li’l Champs Nursery’ writing the same as the home page.
On all the pages, please include a tab at the top with buttons that say ‘home’ ‘about’ ‘childcare’ ‘apply’ and ‘contact us’ all of which are a hyperlink to that said page. 

Remove the ‘gallery’ and the ‘key dates’ hyperlink from the bottom of each page. 

Remove the ‘covid-19 measures’ section on the about page
 - Initial Deployment